move off client to services and make fooditem struct global for later followup pr
